third_party/WebKit/LayoutTests/printing/printing-resize-starts-transition.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<style>
#target {
    padding-top: 1000px;
}

@media screen and (max-width: 600px) {
    #target {
        padding: 0;
        transition: 10s all;
    }
}

@media print {
    #target {
        padding: 0 !important;
    }
}
</style>

<div id="target">
To run this test as a ManualTest:</br>
1. Open this file in Chrome</br>
2. Note that there is a large padding (1000px) at the top of the text</br>
3. Open in Print Preview</br>
4. The test passes if there is no padding at the top.</br>
</div>

<script>
test(function() {
  if (window.internals) {
    assert_equals(getComputedStyle(target).paddingTop, "1000px");
    assert_equals(internals.numberOfPages(200, 200), 1);
    assert_equals(getComputedStyle(target).paddingTop, "1000px");
  }
}, "This test checks that transition are ignored when in @media print");
</script>
